flask-restful-demo/pyproject.toml
[tool.poetry]
name = "flask-restful-demo"
version = "0.1.0"
description = ""
authors = ["liujianwei1 <liujianwei1@tal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask-restful = "^0.3.10"
flask-wtf = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
